requirements.txt
pdftotext
python-dotenv
langchain
openai
tiktoken
click
pinecone-client
flask
flask-sock